Update run_NAMD_GOMC.py
Use relative path to previous directory and correct name for checkpoint file in Config file.  Turned off one instance of the copying, since there is no reason to copy the file locally anymore now that we can use a path.  There may be more instances of where these changes should occur in the script.  Please double check.  I had to do it in the browser editor since this repo contains binaries and I'm on airport wifi.